Implement transformer encoder-decoder
implemented the transformer encoder-decoder with its own functions. (In the future, I want to refactor the code so there is less repetition)

Ran the same experiments as I did before on the GRU (testing generalization as I increase the number of held out names)

The images have incorrect titles because I forgot to change the code.