Refactor code; positional transformer encodingd
Main thing was refactoring the code so the same code works for both the GRU model and the transformer model. To do this, I made a new GRU_Full class.

Another thing I did was add positional encodings and an attention mask to the transformer models.

I also trained two models (one transformer, one GRU) and saved them to the folder. Each were trained for 50 epochs.