Transformer Arithmetic
Fixed some typos in the GRU and Transformer architecture (example- used the decoder pos embedding twice instead of using both the encoder and decoder embeddings).

Also implemented a function which allows us to do transformer arithmetic based on the encoder memory. This was to test if the encoder or the decoder was learning the meaning of the word "herself". We did the arithmetic "Alice compliments herself"-"Alice hugs Emma"+"Bob hugs Emma". If the ENCODER is learning what herself means, then the net output of the encoder is "Bob hugs Alice". If the DECODER is learning what herself means, then the net output is "Bob hugs herself", and we would expect "Bob hugs Bob" to be the decoder's output. Running the experiment on the transformer trained for 100 epochs on the whole corpus, we see that the decoder is the thing learning the representation of the word "herself"