Decoder Only Transformer
Rewrote parts of the old Encoder-Decoder work to create a dataset and model for a decoder-only transformer. The model is as simple as possible: An embedding layer to a 32 hidden dimension, a single self-attention layer (5 heads, each with a head dimension of 8), and a final linear projection to the vocab size. Training for 50 epochs, I am able to get near perfect performance