Encoder/Decoder bug fixed+experiment
Fixed an error where I was not saving the model state dicts correctly. Also set up an experiment where I did every possible transformer arithmetic example to test if a model was encoder/decoder, and found that the model in this case used the decoder to solve the problem 100% of the time.  See ./Experiments with the file name of todays date (070323) to see these results.

I will experiment with training more models tomorrow and see if other models are similar.