Experiments encoder-decoder
Today was the fourth of July, so I just set some experiments up to run in the background.

I ran each of the models I trained before on all possible transformer-arithmetic problems which can determine if the encoder or the decoder is doing the work, and stored the corresponding dictionaries in the appropriately named file in ./Experiments.

Tomorrow I will run a tree decision procedure to determine what parts of the arithmetic sentences are important in determining weather the encoder or decoder is doing the work.